+++ bits/74_xmlbin.js
@@ -28,10 +28,6 @@ function parse_sty(data, name/*:string*/, themes, opts) {
 	return parse_sty_xml((data/*:any*/), themes, opts);
 }
 
-function parse_theme(data/*:string*/, name/*:string*/, opts) {
-	return parse_theme_xml(data, opts);
-}
-
 function parse_sst(data, name/*:string*/, opts)/*:SST*/ {
 	if(name.slice(-4)===".bin") return parse_sst_bin((data/*:any*/), opts);
 	return parse_sst_xml((data/*:any*/), opts);
@@ -55,38 +51,4 @@ function parse_xlink(data, rel, name/*:string*/, opts) {
 function parse_xlmeta(data, name/*:string*/, opts) {
 	if(name.slice(-4)===".bin") return parse_xlmeta_bin((data/*:any*/), name, opts);
 	return parse_xlmeta_xml((data/*:any*/), name, opts);
-}
-
-function write_wb(wb, name/*:string*/, opts) {
-	return (name.slice(-4)===".bin" ? write_wb_bin : write_wb_xml)(wb, opts);
-}
-
-function write_ws(data/*:number*/, name/*:string*/, opts, wb/*:Workbook*/, rels) {
-	return (name.slice(-4)===".bin" ? write_ws_bin : write_ws_xml)(data, opts, wb, rels);
-}
-
-// eslint-disable-next-line no-unused-vars
-function write_cs(data/*:number*/, name/*:string*/, opts, wb/*:Workbook*/, rels) {
-	return (name.slice(-4)===".bin" ? write_cs_bin : write_cs_xml)(data, opts, wb, rels);
-}
-
-function write_sty(data, name/*:string*/, opts) {
-	return (name.slice(-4)===".bin" ? write_sty_bin : write_sty_xml)(data, opts);
-}
-
-function write_sst(data/*:SST*/, name/*:string*/, opts) {
-	return (name.slice(-4)===".bin" ? write_sst_bin : write_sst_xml)(data, opts);
-}
-
-function write_cmnt(data/*:Array<any>*/, name/*:string*/, opts) {
-	return (name.slice(-4)===".bin" ? write_comments_bin : write_comments_xml)(data, opts);
-}
-/*
-function write_cc(data, name:string, opts) {
-	return (name.slice(-4)===".bin" ? write_cc_bin : write_cc_xml)(data, opts);
-}
-*/
-
-function write_xlmeta(name/*:string*/) {
-	return (name.slice(-4)===".bin" ? write_xlmeta_bin : write_xlmeta_xml)();
 }