+++ modules/83_numbers.ts
@@ -830,18 +830,16 @@ function write_numbers_iwa(wb: WorkBook, opts: any): CFB$Container {
 
 	/* build dependent tree */
 	cfb.FileIndex.map((fi, idx): [CFB$Entry, string] => ([fi, cfb.FullPaths[idx]])).forEach(row => {
-		var fi = row[0], fp = row[1];
+		var fi = row[0];
 		if(!fi.name.match(/\.iwa/)) return;
 		var x = parse_iwa_file(decompress_iwa_file(fi.content as Uint8Array));
 
 		x.forEach(ia => {
-			ia.messages.forEach(m => {
-				indices_varint.forEach(ivi => {
-					if(ia.messages.some(mess => varint_to_i32(mess.meta[1][0].data) != 11006 && u8contains(mess.data, ivi[1]))) {
-						dependents[ivi[0]].deps.push(ia.id);
-					}
-				})
-			});
+			indices_varint.forEach(ivi => {
+				if(ia.messages.some(mess => varint_to_i32(mess.meta[1][0].data) != 11006 && u8contains(mess.data, ivi[1]))) {
+					dependents[ivi[0]].deps.push(ia.id);
+				}
+			})
 		});
 	});
 
@@ -1043,10 +1041,10 @@ function write_numbers_iwa(wb: WorkBook, opts: any): CFB$Container {
 							delete tiledata[6]; delete tile[7];
 							var rowload = new Uint8Array(tiledata[5][0].data);
 							tiledata[5] = [];
-							var cnt = 0;
+							//var cnt = 0;
 							for(var R = 0; R <= range.e.r; ++R) {
 								var tilerow = parse_shallow(rowload);
-								cnt += write_tile_row(tilerow, data[R], SST, USE_WIDE_ROWS);
+								/* cnt += */ write_tile_row(tilerow, data[R], SST, USE_WIDE_ROWS);
 								tilerow[1][0].data = write_varint49(R);
 								tiledata[5].push({data: write_shallow(tilerow), type: 2});
 							}